https_server/app/requirements.txt
bottle>=0.12.18
pyzmq>=20.0.0
chardet>=4.0.0